refactor: 디렉토링 명 수정
