--- src/main/java/com/study/petory/domain/dailyQna/dto/response/DailyQuestionGetTodayResponseDto.java
@@ -1,27 +1,0 @@
-package com.study.petory.domain.dailyQna.dto.response;
-
-import com.study.petory.domain.dailyQna.entity.DailyQuestion;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(force = true)
-public class DailyQuestionGetTodayResponseDto {
-
-	private final String question;
-
-	private final String date;
-
-	private DailyQuestionGetTodayResponseDto(String question, String date) {
-		this.question = question;
-		this.date = date;
-	}
-
-	public static DailyQuestionGetTodayResponseDto from(DailyQuestion dailyQuestion) {
-		return new DailyQuestionGetTodayResponseDto(
-			dailyQuestion.getQuestion(),
-			dailyQuestion.getDate()
-		);
-	}
-}
