--- src/main/java/com/study/petory/domain/dailyQna/repository/DailyAnswerRepository.java
@@ -1,8 +1,0 @@
-package com.study.petory.domain.dailyQna.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.study.petory.domain.dailyQna.entity.DailyAnswer;
-
-public interface DailyAnswerRepository extends JpaRepository<DailyAnswer, Long>, DailyAnswerQueryRepository {
-}
