+++ src/main/java/com/study/petory/domain/user/entity/User.java
@@ -7,7 +7,7 @@ import java.util.List;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import com.study.petory.common.entity.TimeFeatureBasedEntity;
 import com.study.petory.domain.album.entity.Album;
-import com.study.petory.domain.dailyQna.entity.DailyAnswer;
+import com.study.petory.domain.dailyqna.entity.DailyAnswer;
 import com.study.petory.domain.pet.entity.Pet;
 import com.study.petory.domain.tradeboard.entity.TradeBoard;
